pyproject.toml
[project]
name = "systemic-risk-index"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black[jupyter]>=25.1.0",
    "fredapi>=0.5.2",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-anthropic>=0.3.17",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.7",
    "langchain-groq>=0.3.6",
    "langchain-openai>=0.3.28",
    "nbformat>=5.10.4",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.0",
    "streamlit>=1.46.1",
    "tabulate>=0.9.0",
    "yfinance>=0.2.65",
]
